chatBot ui test, send messages
