+++ webapp/src/components/chatbot/chat.jsx
@@ -1,4 +1,4 @@
-/* eslint-disable no-unused-vars */
+/* eslint-disable no-unused-varsq */
 import React, { useState } from "react";
 import ChatBot from "react-chatbotify";
 import settings from "./chatSettings"
@@ -59,6 +59,5 @@ async function configure(message) {
         });
     } catch (error) {
         console.error("Error fetching message:", error);
-        return "Error fetching message";
     }
 }
