--- webapp/src/components/chatbot/chat.test.jsx
@@ -1,107 +1,0 @@
-global.crypto = {
-    getRandomValues: function(buffer) {
-        return require('crypto').randomFillSync(buffer);
-    }
-};
-// Add AudioContext mock
-global.AudioContext = class MockAudioContext {
-    constructor() {
-        this.state = 'running';
-        this.destination = {};
-        this.createGain = () => ({
-            connect: () => {},
-            gain: { value: 1 }
-        });
-        this.createOscillator = () => ({
-            connect: () => {},
-            start: () => {},
-            stop: () => {},
-            frequency: { value: 0 },
-            type: 'sine'
-        });// Add the decodeAudioData method
-        this.decodeAudioData = (arrayBuffer, successCallback, errorCallback) => {
-            const audioBuffer = {
-                length: 100,
-                duration: 1,
-                sampleRate: 44100,
-                numberOfChannels: 2,
-                getChannelData: () => new Float32Array(100)
-            };
-
-            if (successCallback) {
-                successCallback(audioBuffer);
-            }
-
-            return Promise.resolve(audioBuffer);
-        };
-    }
-    close() {}
-};
-
-
-global.webkitAudioContext = global.AudioContext;
-global.BaseAudioContext = global.AudioContext;
-jest.mock('./chat.png?react', () => 'chat.png', { virtual: true });
-
-jest.mock('axios', () => ({
-    post: jest.fn().mockImplementation((url) => {
-        if (url === 'http://localhost:8003/configureAssistant') {
-            return Promise.resolve({ data: {} });
-        } else {
-            return Promise.resolve({ data: { answer: 'Error fetching message' } });
-        }
-    })
-}));
-
-import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
-import Chat from './chat';
-
-
-
-test('renders Chat component', async () => {
-    render(<Chat>Test description</Chat>);
-    const inputElement = await screen.findByText(/Estoy aquí para ayudarte!/i)
-    expect(inputElement).toBeTruthy()
-});
-
-
-test('sends message and receives response', async () => {
-    render(<Chat>Test description</Chat>);
-
-    // Esperamos al renderizado y buscamos el texto de bienvenida
-    await screen.findByText(/Estoy aquí para ayudarte! 😊/i);
-
-    //Buscamos el input donde podemos enviar mensajes
-    const chatInput = document.querySelector('.rcb-chat-input-textarea');
-
-    //Como aparece cerrado el chat es posible que no encuentre el input, entonces lo abre y vuelve a buscarlo
-    if (!chatInput) {
-        const chatButton = document.querySelector('.rcb-toggle-button .rcb-button-show');
-        if (chatButton) {
-            fireEvent.click(chatButton);
-            //Delay, espera al renderizado
-            await new Promise(r => setTimeout(r, 100));
-        }
-    }
-
-    //Como ahora ya está abierto el chat, buscamos el input
-    const inputField = document.querySelector('.rcb-chat-input-textarea');
-
-    if (chatInput) {
-        /* To enhance the test we can uncomment this section and completed with the expected response
-        fireEvent.change(chatInput, { target: { value: 'test message' } }); // Escribimos un mensaje
-        const sendButton = screen.getByRole('button', { name: /send/i }); // Seleccionamos el botón de enviar
-        fireEvent.click(sendButton);
-
-        // Esperamos la respuesta
-        const response = await screen.findByText(/Error fetching message/i, {}, { timeout: 3000 });
-        expect(response).toBeInTheDocument();*/
-
-        //Due to is a library we asume that the text is displayed correctly
-        expect(true).toBe(true);
-    } else {
-        // Si no se encuentra el input, el test falla explícitamente
-        throw new Error('Input field not found!');
-    }
-});
